Moving data to CPU in `get_alignn_feats`
Before, one would run into runtime errors such as the following: Expected all tensors to be on the same device, but found at least two devices, cuda:0 and cpu

This would be accompanied by a hard to understand stack trace. I resolve this error by moving data to CPU. This could be improved by having an argument to use CUDA in get_alignn_feats, and then checking if the current version of DGL would allow for CUDA to be used.